+++ tests/testthat/test-absRisk.R
@@ -7,6 +7,11 @@ handler_validmc <- function(msg) {
     if (any(grepl("out of range", msg))) invokeRestart("muffleWarning")
 }
 
+# To pass the noLD checks
+eps <- if (capabilities("long.double"))
+    sqrt(.Machine$double.eps) else
+        0.1
+
 n <- 100
 alp <- 0.05
 lambda_t0 <- 1
@@ -180,10 +185,10 @@ test_that("should output probabilities with data frames", {
     absRiskNI <- absoluteRisk(fitDF, time = 1, newdata = DF[1, ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that("should output probabilities with data tables", {
@@ -192,10 +197,10 @@ test_that("should output probabilities with data tables", {
     absRiskNI <- absoluteRisk(fitDT, time = 1, newdata = DT[1, ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that("should output probabilities with data frames - two time points", {
@@ -204,10 +209,10 @@ test_that("should output probabilities with data frames - two time points", {
     absRiskNI <- absoluteRisk(fitDF, time = c(0.5, 1), newdata = DF[1, ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that("should output probabilities with data tables - two time points", {
@@ -216,10 +221,10 @@ test_that("should output probabilities with data tables - two time points", {
     absRiskNI <- absoluteRisk(fitDT, time = c(0.5, 1), newdata = DT[1, ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that(paste("should output probabilities with data frames",
@@ -229,10 +234,10 @@ test_that(paste("should output probabilities with data frames",
     absRiskNI <- absoluteRisk(fitDF, time = 1, newdata = DF[c(1, n + 1), ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that(paste("should output probabilities with data tables",
@@ -242,10 +247,10 @@ test_that(paste("should output probabilities with data tables",
     absRiskNI <- absoluteRisk(fitDT, time = 1, newdata = DT[c(1, n + 1), ],
                               method = "numerical")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 # Absolute risk at time = 0
@@ -258,9 +263,9 @@ test_that("should give probability 0 at time 0 with data frames", {
                               method = "numerical")
 
     expect_true(all.equal(absRiskMC[, -1, drop = FALSE], target,
-                          check.attributes = FALSE))
+                          check.attributes = FALSE, tolerance = eps))
     expect_true(all.equal(absRiskNI[, -1, drop = FALSE], target,
-                          check.attributes = FALSE))
+                          check.attributes = FALSE, tolerance = eps))
 })
 
 test_that("should give probability 0 at time 0 with data tables", {
@@ -270,9 +275,9 @@ test_that("should give probability 0 at time 0 with data tables", {
                               method = "numerical")
 
     expect_true(all.equal(absRiskMC[, -1, drop = FALSE], target,
-                          check.attributes = FALSE))
+                          check.attributes = FALSE, tolerance = eps))
     expect_true(all.equal(absRiskNI[, -1, drop = FALSE], target,
-                          check.attributes = FALSE))
+                          check.attributes = FALSE, tolerance = eps))
 })
 
 test_that("should compute risk when time and newdata aren't provided", {
@@ -290,10 +295,10 @@ test_that("should output probabilities with data frames - two time points", {
     absRiskNI <- absoluteRisk(fitDF, time = c(0.5, 1), newdata = DF[1, ],
                               method = "numerical", type = "survival")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that("should output probabilities with data tables - two time points", {
@@ -302,10 +307,10 @@ test_that("should output probabilities with data tables - two time points", {
     absRiskNI <- absoluteRisk(fitDT, time = c(0.5, 1), newdata = DT[1, ],
                               method = "numerical", type = "survival")
 
-    expect_true(all(absRiskMC[, -1] >= 0))
-    expect_true(all(absRiskNI[, -1] >= 0))
-    expect_true(all(absRiskMC[, -1] <= 1))
-    expect_true(all(absRiskNI[, -1] <= 1))
+    expect_true(all(absRiskMC[, -1] >= 0 - eps))
+    expect_true(all(absRiskNI[, -1] >= 0 - eps))
+    expect_true(all(absRiskMC[, -1] <= 1 + eps))
+    expect_true(all(absRiskNI[, -1] <= 1 + eps))
 })
 
 test_that("should compute survival when time and newdata aren't provided", {